replace Chat models to Shared library project and ChatHub prototype to server project
