+++ GBSafety.sln
@@ -30,7 +30,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AppModules.ConsoleApp", "sr
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Patterns", "Patterns", "{9A08111E-DDC3-45C6-8F5F-A583B5ABCC0B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PatternsApp", "src\OtherThemes\Patterns\PatternsApp\PatternsApp.csproj", "{58130F9F-7D9B-439D-9038-75830EBE7BF5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PatternsApp", "src\OtherThemes\Patterns\PatternsApp\PatternsApp.csproj", "{58130F9F-7D9B-439D-9038-75830EBE7BF5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PatternsApp.Server", "src\OtherThemes\Patterns\PatternsApp.Server\PatternsApp.Server.csproj", "{318FF33A-3E3A-4671-97A5-3D665C1CE8D9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PatternsApp.Shared", "src\OtherThemes\Patterns\PatternsApp.Shared\PatternsApp.Shared.csproj", "{5DD4C1F6-9603-473B-8D74-52AADE814D9C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -70,6 +74,14 @@ Global
 		{58130F9F-7D9B-439D-9038-75830EBE7BF5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{58130F9F-7D9B-439D-9038-75830EBE7BF5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{58130F9F-7D9B-439D-9038-75830EBE7BF5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{318FF33A-3E3A-4671-97A5-3D665C1CE8D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{318FF33A-3E3A-4671-97A5-3D665C1CE8D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{318FF33A-3E3A-4671-97A5-3D665C1CE8D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{318FF33A-3E3A-4671-97A5-3D665C1CE8D9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5DD4C1F6-9603-473B-8D74-52AADE814D9C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5DD4C1F6-9603-473B-8D74-52AADE814D9C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5DD4C1F6-9603-473B-8D74-52AADE814D9C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5DD4C1F6-9603-473B-8D74-52AADE814D9C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -81,6 +93,8 @@ Global
 		{2158812E-856E-49B2-89FE-DA7F1E099C49} = {90074C60-3E94-41BC-8C48-CA4EFAE76B1F}
 		{9A08111E-DDC3-45C6-8F5F-A583B5ABCC0B} = {8B5B87A5-D495-4D11-AC01-028B3F024071}
 		{58130F9F-7D9B-439D-9038-75830EBE7BF5} = {9A08111E-DDC3-45C6-8F5F-A583B5ABCC0B}
+		{318FF33A-3E3A-4671-97A5-3D665C1CE8D9} = {9A08111E-DDC3-45C6-8F5F-A583B5ABCC0B}
+		{5DD4C1F6-9603-473B-8D74-52AADE814D9C} = {9A08111E-DDC3-45C6-8F5F-A583B5ABCC0B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {5392DE2A-81FB-4D3D-9AA7-61B2AA2636B9}
