+++ src/OtherThemes/Patterns/PatternsApp.Server/Hubs/ChatHub.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.SignalR;
+
+namespace PatternsApp.Server.Hubs;
+
+public class ChatHub : Hub
+{
+    public async Task SendMessage(string user, string message)
+    {
+        await Clients.All.SendAsync("ReceiveMessage", user, message);
+    }
+}
