+++ src/OtherThemes/Patterns/PatternsApp.Shared/ChatInfo.cs
@@ -1,0 +1,7 @@
+﻿namespace PatternsApp.Shared;
+
+public class ChatInfo
+{
+    public Guid Id { get; set; }
+    public string Name { get; set; }
+}
