+++ src/OtherThemes/Patterns/PatternsApp.Shared/ChatsInfo.cs
@@ -1,0 +1,8 @@
+﻿namespace PatternsApp.Shared;
+
+public class ChatsInfo
+{
+    public List<ChatInfo> Chats { get; set; }
+
+    public List<User> Contacts { get; set; }
+}
