+++ src/OtherThemes/Patterns/PatternsApp.Shared/User.cs
@@ -1,0 +1,8 @@
+﻿namespace PatternsApp.Shared;
+
+public class User
+{
+    public Guid UserId { get; set; }
+    public string Color { get; set; }
+    public string Nick { get; set; }
+}
