+++ src/OtherThemes/Patterns/PatternsApp.Shared/UserMessage.cs
@@ -1,0 +1,8 @@
+﻿namespace PatternsApp.Shared;
+
+public class UserMessage
+{
+    public Guid UserId { get; set; }
+    public Guid MessageId { get; set; }
+    public string Text { get; set; }
+}
