+++ src/OtherThemes/Patterns/PatternsApp/Pages/Index.razor
@@ -1,6 +1,4 @@
 ﻿@page "/"
-@using PatternsApp.Models
-
 
 <div>
 	<div class="chats">
